// src/serverApi/v3/models/card-skeleton-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CardSkeletonResponse
 */
export interface CardSkeletonResponse {
    /**
     * 
     * @type {string}
     * @memberof CardSkeletonResponse
     */
    cardId: string;
    /**
     * The approximate height of the referenced card. Intended to be used for prerendering purposes. Note, that different devices can lead to this value not being precise
     * @type {number}
     * @memberof CardSkeletonResponse
     */
    height: number;
}